backend/requirements.txt
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
cryptography
flask
flask-sqlalchemy
flask-login
flask-wtf
flask-limiter
flask-talisman
pyotp
qrcode
pillow